Improve code org
