+++ packages/react-ui/src/widgets/legend/LegendWidgetUI.js
@@ -50,7 +50,7 @@ function LegendWidgetUI({
   collapsed,
   onCollapsedChange,
   onChangeVisibility,
-  onChangeOpacity,
+  onChangeOpacity
 }) {
   const classes = useStyles();
   const isSingle = layers.length === 1;
@@ -62,7 +62,11 @@ function LegendWidgetUI({
         collapsed={collapsed}
         onCollapsedChange={onCollapsedChange}
       >
-        <LegendRows layers={layers} onChangeVisibility={onChangeVisibility} onChangeOpacity={onChangeOpacity} />
+        <LegendRows
+          layers={layers}
+          onChangeVisibility={onChangeVisibility}
+          onChangeOpacity={onChangeOpacity}
+        />
       </LegendContainer>
     </Box>
   );
@@ -167,38 +171,55 @@ const LEGEND_COMPONENT_BY_TYPE = {
 function LegendRows({ layers = [], onChangeVisibility, onChangeOpacity }) {
   const isSingle = layers.length === 1;
 
-  return layers.map(({ id, title, switchable, visible, showOpacityControl = false, opacity = 1, legend = {} }, index) => {
-    const {
-      children = null,
-      type = '',
-      collapsible = true,
-      note = '',
-      attr = ''
-    } = legend;
+  return (
+    <>
+      {layers.map(
+        (
+          {
+            id,
+            title,
+            switchable,
+            visible,
+            showOpacityControl = false,
+            opacity = 1,
+            legend = {}
+          },
+          index
+        ) => {
+          const {
+            children = null,
+            type = '',
+            collapsible = true,
+            note = '',
+            attr = ''
+          } = legend;
 
-    const isLast = layers.length - 1 === index;
-    // TODO: Add validation for layer.type
-    const hasChildren = LEGEND_COMPONENT_BY_TYPE[type] || children;
-    const LegendComponent = LEGEND_COMPONENT_BY_TYPE[type] || (() => children);
-    return (
-      <Fragment key={id}>
-        <LegendWrapper
-          id={id}
-          title={title}
-          collapsible={!!(collapsible && hasChildren)}
-          switchable={switchable}
-          visible={visible}
-          note={note}
-          attr={attr}
-          showOpacityControl={showOpacityControl}
-          opacity={opacity}
-          onChangeOpacity={onChangeOpacity}
-          onChangeVisibility={onChangeVisibility}
-        >
-          <LegendComponent legend={legend} />
-        </LegendWrapper>
-        {!isSingle && !isLast && <Divider />}
-      </Fragment>
-    );
-  });
+          const isLast = layers.length - 1 === index;
+          // TODO: Add validation for layer.type
+          const hasChildren = LEGEND_COMPONENT_BY_TYPE[type] || children;
+          const LegendComponent = LEGEND_COMPONENT_BY_TYPE[type] || (() => children);
+          return (
+            <Fragment key={id}>
+              <LegendWrapper
+                id={id}
+                title={title}
+                collapsible={!!(collapsible && hasChildren)}
+                switchable={switchable}
+                visible={visible}
+                note={note}
+                attr={attr}
+                showOpacityControl={showOpacityControl}
+                opacity={opacity}
+                onChangeOpacity={onChangeOpacity}
+                onChangeVisibility={onChangeVisibility}
+              >
+                <LegendComponent legend={legend} />
+              </LegendWrapper>
+              {!isSingle && !isLast && <Divider />}
+            </Fragment>
+          );
+        }
+      )}
+    </>
+  );
 }
